PAAC-668: Code to fix 5,000,000,000 error message for pre-2015 Input Fields,

+++ app/form/Pre2015DefinedBenefitForm.scala
@@ -24,14 +24,14 @@ import config.AppSettings
 case class Pre2015DefinedBenefitModel(definedBenefit: BigDecimal) extends Pre2015Fields
 
 trait Pre2015DefinedBenefitForm extends Pre2015FormFactory {
-  def apply(year: Int): Form[_ <: Pre2015Fields] = {
+  def apply(year: Long): Form[_ <: Pre2015Fields] = {
     if (isPoundsAndPence)
       Form[Pre2015DefinedBenefitModel](mapping(penceDB(year))(Pre2015DefinedBenefitModel.apply)(Pre2015DefinedBenefitModel.unapply))
     else
       Form[Pre2015DefinedBenefitModel](mapping(poundsDB(year))(toModel)(Pre2015DefinedBenefitModel.unapply))
   }
 
-  protected val toModel: (Int) => Pre2015DefinedBenefitModel = (i: Int) =>
+  protected val toModel: (Long) => Pre2015DefinedBenefitModel = (i: Long) =>
     Pre2015DefinedBenefitModel(BigDecimal(i))
 }
